+++ src/subscriber.ts
@@ -17,20 +17,6 @@ export default class ConnectionSubscriber extends ConnectionBase {
    */
   // biome-ignore lint/suspicious/noConfusingVoidType: stream が <MediaStream | void> なのでどうしようもない
   async connect(): Promise<MediaStream | void> {
-    // options.multistream が明示的に false を指定した時だけレガシーストリームにする
-    if (this.options.multistream === false) {
-      const stream = await Promise.race([
-        this.legacyStream().finally(() => {
-          this.clearConnectionTimeout()
-          this.clearMonitorSignalingWebSocketEvent()
-        }),
-        this.setConnectionTimeout(),
-        this.monitorSignalingWebSocketEvent(),
-      ])
-      this.monitorWebSocketEvent()
-      this.monitorPeerConnectionState()
-      return stream
-    }
     await Promise.race([
       this.multiStream().finally(() => {
         this.clearConnectionTimeout()
@@ -41,61 +27,6 @@ export default class ConnectionSubscriber extends ConnectionBase {
     ])
     this.monitorWebSocketEvent()
     this.monitorPeerConnectionState()
-  }
-
-  /**
-   * レガシーストリームで Sora へ接続するメソッド
-   *
-   * @deprecated この関数は非推奨です、マルチストリームを利用してください。
-   *
-   */
-  private async legacyStream(): Promise<MediaStream> {
-    await this.disconnect()
-    const ws = await this.getSignalingWebSocket(this.signalingUrlCandidates)
-    const signalingMessage = await this.signaling(ws)
-    await this.connectPeerConnection(signalingMessage)
-    if (this.pc) {
-      this.pc.ontrack = async (event): Promise<void> => {
-        this.stream = event.streams[0]
-        const streamId = this.stream.id
-        if (streamId === 'default') {
-          return
-        }
-        const data = {
-          'stream.id': streamId,
-          id: event.track.id,
-          label: event.track.label,
-          enabled: event.track.enabled,
-          kind: event.track.kind,
-          muted: event.track.muted,
-          readyState: event.track.readyState,
-        }
-        this.writePeerConnectionTimelineLog('ontrack', data)
-        this.callbacks.track(event)
-        this.stream.onremovetrack = (event): void => {
-          this.callbacks.removetrack(event)
-          if (event.target) {
-            const streamId = (event.target as MediaStreamTrack).id
-            const index = this.remoteConnectionIds.indexOf(streamId)
-            if (-1 < index) {
-              delete this.remoteConnectionIds[index]
-            }
-          }
-        }
-        if (-1 < this.remoteConnectionIds.indexOf(streamId)) {
-          return
-        }
-        this.remoteConnectionIds.push(streamId)
-      }
-    }
-    await this.setRemoteDescription(signalingMessage)
-    await this.createAnswer(signalingMessage)
-    this.sendAnswer()
-    if (this.pc?.getConfiguration().iceTransportPolicy === 'all') {
-      await this.onIceCandidate()
-    }
-    await this.waitChangeConnectionStateConnected()
-    return this.stream || new MediaStream()
   }
 
   /**